feat: derive more Route props on instantiation
fixes #1423

There are a few things going on in this commit. First and foremost, I've
ensured that several route properties (id, path, fullPath, parentRoute,
to) are actually created inside the constructor so they are available
immediately after instantiation.

The logic for creating these properties was involved enough that I
thought it was useful to keep it in a separate function. However, I did
opt to return the properties from the function so that they can still be
set on the instance inside the constructor. That allows use to not lie
to TypeScript about the availability of these properties. (So I removed
the `!` from the type declarations.)

Finally, I simplified a lot of the logic that was used to set all the
properties themselves. By declaring the variables early, then setting
all of them inside a single if/else block based on if we are in the root
route or not, you can see much more clearly the difference between how
the root route properties are created and how all other route properties
are created. There were also several places where a value was set on a
variable, used to derive a separate variable, then re-set again to
finalize the original variable. That just isn't necessary when you
branch once based on whether or not you are in the root route.