fix(history): when hash history is used with search params
When search params where used with hash history then the search
parameters where not read nore writen to the correct location in the
href. The correct location for search params according to [RFC3986](https://www.rfc-editor.org/rfc/rfc3986#section-3) is:

URI = scheme ":" hier-part [ "?" query ] [ "#" fragment ]

Those are the parameters found in `window.location.search` but the hash
history read and wrote them here:

URI = scheme ":" hier-part [ "#" fragment ] [ "?" query ]

To fix that we need to account for the search parameters when
transforming the href to and from the hash route url.

closes #1449

BREAKING CHANGE: Any website that relies on the previous none RFC 3986
compatible format and use that format when linking to a page with search
parameters and a hash route will no longer get their search parameters
parsed.