fix(history) make hash history behaviour backwards compatible
When changing hash history to respect search params according to [https://www.rfc-editor.org/rfc/rfc3986#section-3](rfc3986)
a breaking change was introduced. Links to websites using react-router with
query parameters and hash history will not work as before. To fix that
we merge search parameters found in the hash fragment with those found
in the query string. This should allow backwards compatibility and avoid
breaking changes. However it is also a bit hacky and not the expected
behaviour.