fix: custom React.lazy
Apparently React.lazy will always suspend, even if the component is synchronously available. Makes sense, but this is annoying when used with our code-splitting paradigm. Since we preload the underlying components and even do our own top-level transition suspense for these components, it doesn't make sense to cause a React suspense boundary to suspend even if we have surely preloaded the component.

With this fix, if the component has been preloaded, it will synchronously render, otherwise it will suspend until it's been loaded.