fix(react-router): do not eagerly garbage collect routes (#1619)
When the user sets a `defaultGcTime` on the router or a `gcTime` on a route to the value of `0`, the current eager garbage collection we've got going on ends up setting the route-context to `undefined`.

This normally wouldn't matter on direct navigation, however, when redirects come into play this results in the route trying to execute its options (`beforeLoad`, `loader`, `component`, etc) without its processed route context being available. The current setup looks to perform garbage collection both before and after a `this.loadMatches` has been called in `router.load`.

This change makes it so that garbage collection is done after the `this.loadMatches` call has resolved.