feat: wrap "use server" over callexprs/identifiers for createServerFn
Wrap identifiers/call expressions that are passed as the 2nd parameter
to createServerFn with a function expression whose body is prepended
with the "use server" directive.

This allows for functions that are passed to createServerFn to be
wrapped with other functions that server as middleware.

router-vite-plugin/tests/test-files/createServerFn.tsx has a Zod
middleware function example which has the "use server" directive
properly prepended.