fix: only disallow preload match updates to active matches **at the time of preload**
This fixes a special case where you could hover over a link and start a preload, then click the link before it resolved. Before, any preload updates (including a successful state update) were ignored if the match was active, so in this case, the newly active match that hadn't resolved yet would never get updated since it is now active.

This change only disallows updates to the original match IDs that are currently active.