fix(start): load meta only on successful hydration
When a route is matched to be unsuccessful such that no loader data is
available, the meta function still gets called with undefined loader
data. Given that the type signature of the meta function assumes that
loader data is always to be defined, a change was made to only call the
meta function when a match is successful.