Determine validity of standard schema result based on existence of issues, instead of value
Some implementations (namely Valibot) still include the value key for failure results, which is technically still correct in structural typing (since an object type is a *minimum* requirement). What can actually be relied on is that result.issues is undefined if the result is valid, as this is in the spec.

Currently failure results from Valibot are still treated as success.