fix(solid-router): make HeadContent work for client navigation (#3616)
Picking solid-router relevant changes from
[commit](https://github.com/TanStack/router/pull/3522/commits/9eaa09ea1f95f786e9bd239fe787f10d6cb08e1c)
by @brenelz , that fixes assets handler in solid-router by using
`@solidjs/meta`.

With this, the head tags like the title will update on client
navigation.

The changes in this PR that affect most files are merely repetitions of:
- bumping solid 1.9.4 -> 1.9.5
- inserting `<HeadContent>` in the root route
- removing `<title>` from the index.html

Includes an automated test for solid-router corresponding to 
- https://github.com/TanStack/router/pull/3610

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>