fix(react-router,solid-router): scroll to element does not work for hash history (#3700)
The `restoreScroll` function was directly interfacing with
`window.location.hash` to get the 'hash' value, which conflicts with how
hash routing functions. As such, when the hash value was retrieved, it
came back as `#/page#foo` instead of just `#foo` like it would with
regular browser history.

Switching it to use the `history` driver attached to the `router`
instance, correctly returns the `hash` value irrespective of the type of
routing being used.

Fixes #3479