docs: fix create-tsrouter-app command (#3829)
I ran into errors trying to use the npm and yarn commands listed here,
so I looked at https://github.com/TanStack/create-tsrouter-app and it
appears you should use npx with the `--package-manager` option to
specify npm/yarn/etc.

I see the doc was updated in
https://github.com/TanStack/router/pull/3759/files#diff-32b91344fa0565552fb394a3166cbdebd81afa9d02ffec1ed2c1d055ebad46a9,
I assume that was the old way to create a project but feel free to edit
or close this PR if there's a better way to document this, thanks.