fix(router-plugin): code-splitting to handle default exported function declarations (#3901)
Currently, when the code-splitter encounters a known identifier used for
configuration in `createFileRoute` (such as `component`, `loader`,
`errorComponent`, etc.) it retains the export without splitting it out
into a separate file.

This functionality is currently partially working when the export is a
"default export" which references an already declared identifier.

```ts
const loader = () => {}
export default loader
```

However, it does not work when the "default export" is also making a
function declaration for the export.

```ts
export default function loader() {}
```

This PR adds the functionality to detect this use case and perform the
same transformations to align itself with the behaviour for all others
exports.