fix(react,solid): extract CSS for preloading
the vite manifest has a key named `css` which stores
the locations of the related CSS files tied to the
individual vite build artifacts.

this changes pulls those artifacts out at build time
and imports it during hydration.

this change works for both React and Solid. Solid has
a issue where deduplication isn't working at the
moment, but this change does not cause it so it should
be investigated separately