fix(router-plugin): avoid blocking webpack build command
Currently a chokidar watcher is registered outside production mode of
webpack. This is okay for serve and watch commands, but unwanted for
the build command.

A webpack build command run should process one compilation and shutdown.
This is not the case as the plugin is blocked by the chokidar watcher.

To alleviate this limit the chokidar watcher only when webpack is
running with watchMode enabled. The easiest way to accomplish this is to
register the watcher within the webpack watchRun hook.

Signed-off-by: Valentin Kunz <vale.kunz@googlemail.com>