fix: declare `vite` as a `peerDependency` (#4135)
Currently, none of the packages (i.e `packages/*`) need `vite`
bundled as a `dependency`. So its been moved to the
`devDependencies` and had the version opened using
`"peerDependencies": { "vite": ">=6.0.0" }`.

Additionally, `vite` has been added as a `dependency` to
the Start end-to-end sandboxes.

In a later PR, `vite` will need to be added to all the
TanStack Start examples (i.e. `examples/**/start-*`).