fix: resolve custom `client` and `server` entry paths (#4172)
This PR fixes it, so that a user can define a custom framework entry for both
the `client` and `server` (in either `src/client.tsx` or `src/server.tsx`).

This PR also removes the unneeded `src/server.ts` files present as part of the
`devinxi` migration.

Additionally, the React client entry has been change to wrap `hydrateRoot`
with `startTransition`. This change useful, since after the SSR-rendered
markup is sent to the client, `startTransition` lets React know to defer the
hydration in favor of user events.

fixes #4138