feat(server-functions-plugin): make the manifest output path configurable (#4232)
As part of the devinxi changes, the following hard-coded output
path of the server functions manifest file was changed.

This would have had breaking changes in the user DX for downstream
frameworks building upon the `@tanstack/server-functions-plugin`
package - (i.e. SolidStart).

This change makes the output path of the manifest file for the
`server-functions-plugin` package, configurable.