fix(solid-router): loaders being called twice during SSR (#4574)
Basically my understanding is that the `load` function gets called once
during SSR but right now our `Transitioner` calls it again as it appears
`createRenderEffect` run on the server too where as the `useEffect` in
React doesn't.

The fix is just to bail out if its running on the server. Feels kinda
hacky but it works.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>