fix: set component return type to `any` by default
TypeScript cannot infer the return type of a component if it directly (e.g. inside a fragment) returns route state.

```tsx
const fooRoute = createRoute({
  getParentRoute: () => rootRoute,
  path: 'foo/$id',
  validateSearch: (search) => ({ hello: search.hello }) as { hello: string },
  component: () => <>foo {fooRoute.useParams().id}</>,
})
```

This results in the following compiler error:

> 'component' implicitly has return type 'any' because it does not have a return type annotation and is referenced directly or indirectly in one of its return expressions.ts(7023)

This commit reverts the change in #4528 that required the return type of a route component to be of `React.ReactNode` and uses `any` again.

Strict typing of the route components can be enabled via module augmentation:

```tsx
declare module '@tanstack/react-router' {
  interface RouteTypes<TProps> {
    component: React.FC<TProps>
  }
}
```

In the strict case, the above example compiles with an explicit return type annotation:

```tsx
const fooRoute = createRoute({
  getParentRoute: () => rootRoute,
  path: 'foo/$id',
  validateSearch: (search) => ({ hello: search.hello }) as { hello: string },
  component: () : React.JSX.Element => <>foo {fooRoute.useParams().id}</>,
})
```

fixes #4560