fix: invalid path generation for params above groups routes (#4629)
Fixes #4599

# Previous behavior

```
$param
    (group)
        route.tsx
        index.tsx
```

when router-generator processes the `route.tsx` node, the `split`
variable


https://github.com/TanStack/router/blob/8df504c3a46390a1331fffd72fcfec52bdb5a162/packages/router-generator/src/generator.ts#L1258

will be `[ '$param', '(group)' ]`, and because the last part is a group


https://github.com/TanStack/router/blob/8df504c3a46390a1331fffd72fcfec52bdb5a162/packages/router-generator/src/generator.ts#L1263

the whole route will be generated as pathless, i.e.

```ts
const IndexRoute = IndexRouteImport.update({
  id: '/',
  path: '/',
  getParentRoute: () => rootRouteImport,
} as any)
const ParamgroupRouteRoute = ParamgroupRouteRouteImport.update({
  id: '/$param/(group)',
  ////////////////
  path: '/$param/', // <---- This will NOT be here without the fix
  ////////////////
  getParentRoute: () => rootRouteImport,
} as any)
const ParamgroupIndexRoute = ParamgroupIndexRouteImport.update({
  id: '/',
  path: '/',
  getParentRoute: () => ParamgroupRouteRoute,
} as any)
```

therefore the `$param` part of the path will be lost, and it won't be
included in `index.tsx`'s path resolution algorithm (in this case its
path would be `/`)

# Current behavior

When processing a route node, its path is skipped when all nodes between
it and the higher route are groups.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>