feat(router-core) path: parsePathname uses integers for Segment type (#4711)
Using integers instead of strings when possible is always better for
performance. It uses less memory, and enables faster comparisons.

This is also preparatory work for a build-time `parsePathname` as it
will result in a smaller output (though once gzipped it might not make a
difference in downloaded kB in the end).