fix(router-core): buildLocation must consider basePath when resolving fromPath (#4930)
when building the location we are already considering the `basepath` for
the `to` path resolution but we don't do the same for the `from` path.

This PR resolves the `from` path also with the `basepath` so as to
ensure a like for like comparison with `to`.

This PR resolves #4924 and also adds a test to check for this regression
going forward

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>