docs: Use named component in react quick-start
When using an anonymous component, Vite's hot reloading doesn't always immediately pick up changes made to the component. This can be confusing to someone testing out the router.

By extracting a named component, Vite seems to be able to match updates to the component to the mounted instance and apply hot reloading as expected.